refactor: move pemfilereader to data mod
