+++ src-tauri/src/data/config.rs
@@ -1,0 +1,7 @@
+use super::{path::CONFIG, PemfileReader};
+
+pub fn pem_reader() -> PemfileReader {
+  let mut filepath = CONFIG.clone();
+  filepath.push("identity.pem");
+  PemfileReader { filepath }
+}
