+++ src-tauri/src/data/mod.rs
@@ -1,7 +1,10 @@
+mod config;
 mod path;
+mod pem_reader;
 pub mod sqlite;
 pub mod tauri;
 mod usr_ident;
 mod usr_info;
 pub use usr_ident::*;
 pub use usr_info::*;
+pub use pem_reader::PemfileReader;
