+++ src-tauri/src/network/p2p_tunl/mod.rs
@@ -1,5 +1,4 @@
 mod p2p_loop;
-mod pem_reader;
 mod resolver;
 pub use p2p_loop::p2p_loop;
 pub mod signal;
